requirements.txt
asgiref==3.11.0
Django==5.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg2-binary==2.9.11
PyJWT==2.10.1
sqlparse==0.5.5
tzdata==2025.3
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.3.0
python-dotenv==1.0.1
openpyxl==3.1.2
